cpuexec.h: Deduplicate S9xCheckInterrupts in the output executable.
This code does not modify the timing of any SNES instruction.

S9xCheckInterrupts is called only in situations where the CPU involved is the main 65C816 CPU, so having multiple distinct implementations containing the same code when they do not apply to the SA-1 CPU places a greater burden on the host's instruction cache.