tile.cpp: Preload GFX.ScreenColors[0] into a variable in backdrop renderers.
The value of GFX.ScreenColors[0] is loaded conditionally. Without this manual preloading, the compiler cannot move the load out of the render loop. Because it's a conditional pointer access, moving it out of the loop could, according to the C standard, cause a NULL dereference that cannot be made unconditional if it was conditional.