tile.cpp: Perform computations to draw 2 pixels at once.
CANNOT BE SENT UPSTREAM, due to the code needed to support SNES hi-res mode not being there.

On MIPS, memory loads for multiple pixels may be launched concurrently because there are enough registers to hold all of them. It is advantageous to load more because, if a register's value is needed immediately after it is loaded from memory and it is not loaded yet, the program waits, doing nothing.